Move 'service' query to layout component
