+++ web/src/containers/layout.js
@@ -1,45 +1,28 @@
 import { graphql, StaticQuery } from 'gatsby';
 import React from 'react';
-// import React, { useState } from 'react';
 import Layout from '../components/layout';
 
-const query = graphql`
-  query SiteTitleQuery {
-    site: sanitySiteSettings(_id: { regex: "/(drafts.|)siteSettings/" }) {
-      title
+export const query = graphql`
+  query LayoutQuery {
+    services: allSanityService {
+      nodes {
+        title
+        slug {
+          current
+        }
+      }
     }
   }
 `;
 
-function LayoutContainer(props) {
-  // const [showNav, setShowNav] = useState(false);
-  // function handleShowNav() {
-  //   setShowNav(true);
-  // }
-  // function handleHideNav() {
-  //   setShowNav(false);
-  // }
-  return (
-    <StaticQuery
-      query={query}
-      render={(data) => {
-        if (!data.site) {
-          throw new Error(
-            'Missing "Site settings". Open the studio at http://localhost:3333 and add "Site settings" data',
-          );
-        }
-        return (
-          <Layout
-            {...props}
-            // showNav={showNav}
-            // siteTitle={data.site.title}
-            // onHideNav={handleHideNav}
-            // onShowNav={handleShowNav}
-          />
-        );
-      }}
-    />
-  );
-}
+const LayoutContainer = (props) => (
+  <StaticQuery
+    query={query}
+    render={(data) => {
+      const services = (data || {}).services.nodes;
+      return <Layout {...props} services={services} />;
+    }}
+  />
+);
 
 export default LayoutContainer;
