+++ web/src/pages/index.js
@@ -1,80 +1,10 @@
 import React from 'react';
-import { graphql } from 'gatsby';
-// import {
-//   mapEdgesToNodes,
-//   filterOutDocsWithoutSlugs,
-//   filterOutDocsPublishedInTheFuture,
-// } from '../lib/helpers';
-// import Container from '../components/container';
 import GraphQLErrorList from '../components/graphql-error-list';
-// import ProjectPreviewGrid from '../components/project-preview-grid';
 // import SEO from '../components/seo';
 import Layout from '../containers/layout';
 
-export const query = graphql`
-  query IndexPageQuery {
-    services: allSanityService {
-      nodes {
-        title
-        slug {
-          current
-        }
-      }
-    }
-  }
-`;
-
-// export const query = graphql`
-//   query IndexPageQuery {
-//     site: sanitySiteSettings(_id: { regex: "/(drafts.|)siteSettings/" }) {
-//       title
-//       subtitle
-//       description
-//       keywords
-//     }
-//     projects: allSanityProject(
-//       limit: 6
-//       sort: { fields: [title], order: DESC }
-//       filter: { slug: { current: { ne: null } }}
-//     ) {
-//       edges {
-//         node {
-//           id
-//           mainImage {
-//             crop {
-//               _key
-//               _type
-//               top
-//               bottom
-//               left
-//               right
-//             }
-//             hotspot {
-//               _key
-//               _type
-//               x
-//               y
-//               height
-//               width
-//             }
-//             asset {
-//               _id
-//             }
-//             alt
-//           }
-//           title
-//           _rawExcerpt
-//           slug {
-//             current
-//           }
-//         }
-//       }
-//     }
-//   }
-// `;
-
 const IndexPage = (props) => {
-  const { data, errors } = props;
+  const { errors } = props;
 
   if (errors) {
     return (
@@ -84,10 +14,8 @@ const IndexPage = (props) => {
     );
   }
 
-  const services = (data || {}).services.nodes;
-
   return (
-    <Layout services={services}>
+    <Layout>
       {/* <SEO title="Whitten Associates" />
       <Container>
         <h1>Welcome to Whitten Associates</h1>
