+++ web/src/pages/portfolio.js
@@ -1,23 +1,9 @@
 import React from 'react';
-import { graphql } from 'gatsby';
 import GraphQLErrorList from '../components/graphql-error-list';
 import Layout from '../containers/layout';
 
-export const query = graphql`
-  query PortfolioPageQuery {
-    services: allSanityService {
-      nodes {
-        title
-        slug {
-          current
-        }
-      }
-    }
-  }
-`;
-
 const PortfolioPage = (props) => {
-  const { data, errors } = props;
+  const { errors } = props;
   if (errors) {
     return (
       <Layout>
@@ -26,10 +12,8 @@ const PortfolioPage = (props) => {
     );
   }
 
-  const services = (data || {}).services.nodes;
-  
   return (
-    <Layout services={services} headline="Portfolio">
+    <Layout headline="Portfolio">
       <div>Here's a Portfolio:</div>
       <div>[pictures of buildings]</div>
     </Layout>
